requirements.txt
# Data pipeline
pandas>=2.0
matplotlib>=3.8
pyarrow>=14.0
python-dotenv>=1.0
gdown>=5.0
scikit-learn>=1.4
joblib>=1.3

# Schema validation
pydantic>=2.0

# API (uncomment when src/api.py is ready)
# fastapi>=0.110
# uvicorn[standard]>=0.29
